+++ gulp/tasks/default.js
@@ -1,6 +1,7 @@
 import gulp from 'gulp'
 import gulpSequence from 'gulp-sequence'
 import { getTasks } from '../libs/utils'
+import util from 'gulp-util'
 
 export default function defaultTask(cb) {
 	const { assetTasks, codeTasks } = getTasks()