+++ src/docs/tokens/tokens.config.js
@@ -2,6 +2,10 @@ const path = require('path')
 module.exports = {
 	context: {
 		colors: require(path.join(process.cwd(), 'src/tokens/colors.json')),
-		fonts: require(path.join(process.cwd(), 'src/tokens/fonts.json'))
+		fonts: require(path.join(process.cwd(), 'src/tokens/fonts.json')),
+		breakpoints: require(path.join(
+			process.cwd(),
+			'src/tokens/breakpoints.json'
+		))
 	}
 }