+++ CHANGELOG.md
@@ -3,6 +3,15 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+### 0.1.5
+
+#### Added
+
+- Integration with craftcms
+- Better path config handling
+- Generate static builds from fractal library
+- Critical css all the things
+
 ### 0.1.4
 
 #### Update